requirements.txt
pdfplumber
python-pptx
thefuzz
python-Levenshtein
google-generativeai
python-dotenv
openpyxl
python-docx
